simplest version in working order
The feather board has a "USB" pin on it that delivers 5V+, which is all I needed to power this single fan. That allowed me to remove the PowerBoost 500 Basic that I had in between.

When I had a separate power supply for most of this work, the ground of the feather was not shared with the ground of everything else. Remove the extra power supply fixed this issue, but I'm still not sure the proper way to get around that in the future.